+++ ChangeLog.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+- Support node version 8 and protocol version 8.
+- Support for suspend/resume in validator update transactions.
+- Add command `consensus detailed-status` for getting detailed consensus status (from protocol
+  version 6).
+- Add `raw GetConsensusDetailedStatus` that presents the detailed consensus status as JSON.
+- Update GHC version to 9.6.6 (lts-22.39).
+
 ## 7.0.1
 
 - Display the correct "at disposal" balance when used with node versions prior to 7.